client/requirements.txt
pandas~=2.2.2
numpy~=1.26.4
scikit-learn~=1.5.1
sqlalchemy~=2.0.25
datetime~=5.5
pydantic~=2.5.3
fastapi~=0.103.0
starlette~=0.27.0
faker~=25.6.0
scapy~=2.4.3
passlib~=1.7.4
psutil~=5.9.8
websockets~=12.0
uvicorn~=0.20.0
watchdog
pcapy-ng
